# findproject/docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'FindThePlace'
  version: '0.0.1'
servers:
  - url: http://localhost:8000/api/v1/
    description: dev server
paths:
...
  /users/{id}:
    security:
      - Token: []
    get:
      operationId: Профиль пользователя
      description: 'Доступно всем авторизованным пользователям.'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: statistic
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithStatistic'
          description: ''
      tags:
        - Пользователи
...
  /users/me/avatar:
    security:
      - Token: []
    put:
      operationId: Добавление аватара
      description: 'Добавление аватара текущего пользователя (Base64)'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAvatar'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetAvatarResponse'
          description: 'Аватар успешно добавлен'
      tags:
        - Пользователи
    delete:
      operationId: Удаление аватара
      description: 'Удаление аватара текущего пользователя'
      responses:
        '204':
          description: 'Аватар успешно удален'
      tags:
        - Пользователи

  # ==================== ДРУЗЬЯ ====================
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Уникальный юзернейм"
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        avatar:
          type: string
          format: uri
          nullable: true
        avatar_url:
          type: string
          format: uri
          nullable: true
        is_friend:
          type: boolean
          readOnly: true
    UserWithStatistic:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            routes:
              type: array
              items:
                $ref: '#/components/schemas/Route'
            routes_count:
              type: integer
...
    SetAvatar:
      type: object
      required:
        - avatar
      properties:
        avatar:
          type: string
          description: 'Изображение в формате Base64'
    SetAvatarResponse:
      type: object
      properties:
        avatar:
          type: string
          format: uri

    # ==================== ЛОКАЦИИ ====================
...
    Location:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        radius:
          type: number
          format: float
          description: "Радиус поиска (метры)"
          nullable: true
        address:
          type: string
          maxLength: 250
        time:
          type: string
          format: date-time
          description: "Время на точке"

    # ==================== МАРШРУТЫ ====================
    Route:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        start:
          $ref: '#/components/schemas/Location'
        end:
          $ref: '#/components/schemas/Location'
        distance:
          type: number
          format: float
          description: "Расстояние (метры)"
        time:
          type: string
          format: time
          description: "Время прохождения (HH:MM:SS)"
          example: "01:30:00"
        date:
          type: string
          format: date
          description: "Дата прохождения"
          example: "2024-01-20"
...
  securitySchemes:
    Token:
      description: 'Авторизация по токену. Используйте заголовок "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token